[cpp] Split 10x and AnnData import code files
As a side-effect, it seems the SingletonNumReader
was dead code and got deleted.
