+++ r/src/Makevars.in
@@ -35,7 +35,8 @@ bpcells-cpp/fragmentIterators/StoredFragments.o \
 bpcells-cpp/fragmentUtils/BedWriter.o \
 bpcells-cpp/fragmentUtils/FootprintMatrix.o \
 bpcells-cpp/fragmentUtils/InsertionIterator.o \
-bpcells-cpp/matrixIterators/ImportMatrixHDF5.o \
+bpcells-cpp/matrixIterators/ImportMatrix10xHDF5.o \
+bpcells-cpp/matrixIterators/ImportMatrixAnnDataHDF5.o \
 bpcells-cpp/matrixIterators/PeakMatrix.o \
 bpcells-cpp/matrixIterators/TileMatrix.o \
 bpcells-cpp/matrixIterators/MatrixStats.o \
